Start converting spaces to tabs, with explanation
Until recently I was 'spaces over tabs' developer, but I recently read
[a blog post by Bruce Wray about the accessibility argument for tabs over spaces](https://www.brycewray.com/posts/2022/06/accessibility-argument-tabs-spaces/)
and followed the links he gave, and the ability of tabs to be handled by braille
or screen reader devices (or users who need to zoom their displays) is more than
just 'taste' but can in fact be a huge win for 'getting stuff done'. That major
sites don't handle tabs well in code blocks is a reflection on poor site design,
and should not be excused or catered to.

All it takes is (for example)

``` css
code,
pre {
	tab-size: 3;
}
```

And of course, you can use CSS selectors to override when needed.